+++ src/test/java/com/salesforce/dynamodbv2/mt/mappers/MtAmazonDynamoDbStreamsByTableTest.java
@@ -22,7 +22,8 @@ import org.junit.jupiter.api.Test;
 class MtAmazonDynamoDbStreamsByTableTest {
 
     private static final String TABLE_PREFIX = MtAmazonDynamoDbStreamsByTableTest.class.getSimpleName() + ".";
-    private static MtAmazonDynamoDbStreamsBaseTestUtils BASE_TEST_UTILS = new MtAmazonDynamoDbStreamsBaseTestUtils();
+    private static final MtAmazonDynamoDbStreamsBaseTestUtils BASE_TEST_UTILS
+        = new MtAmazonDynamoDbStreamsBaseTestUtils();
 
     // work-around for command-line build: some previous tests don't seem to be clearing the mt context
     @BeforeEach
@@ -45,7 +46,8 @@ class MtAmazonDynamoDbStreamsByTableTest {
             .withContext(MT_CONTEXT)
             .build();
         try {
-            TableUtils.createTableIfNotExists(dynamoDb, BASE_TEST_UTILS.newCreateTableRequest(randomTableName, false));
+            TableUtils.createTableIfNotExists(dynamoDb,
+                MtAmazonDynamoDbStreamsBaseTestUtils.newCreateTableRequest(randomTableName, false));
 
             MtAmazonDynamoDbStreams mtDynamoDbStreams = MtAmazonDynamoDbStreams.createFromDynamo(mtDynamoDb,
                 AmazonDynamoDbLocal.getAmazonDynamoDbStreamsLocal());
@@ -54,13 +56,13 @@ class MtAmazonDynamoDbStreamsByTableTest {
             List<Stream> streams = mtDynamoDbStreams.listStreams(new ListStreamsRequest()).getStreams();
             assertTrue(streams.isEmpty());
 
-            BASE_TEST_UTILS.createTenantTables(mtDynamoDb);
+            MtAmazonDynamoDbStreamsBaseTestUtils.createTenantTables(mtDynamoDb);
 
             // getting streams after creating tenant tables should return as many streams as tenants
             streams = mtDynamoDbStreams.listStreams(new ListStreamsRequest()).getStreams();
-            assertEquals(BASE_TEST_UTILS.TENANTS.length, streams.size());
+            assertEquals(MtAmazonDynamoDbStreamsBaseTestUtils.TENANTS.length, streams.size());
         } finally {
-            BASE_TEST_UTILS.deleteMtTables(mtDynamoDb);
+            MtAmazonDynamoDbStreamsBaseTestUtils.deleteMtTables(mtDynamoDb);
             TableUtils.deleteTableIfExists(dynamoDb, new DeleteTableRequest(randomTableName));
         }
     }
@@ -78,14 +80,18 @@ class MtAmazonDynamoDbStreamsByTableTest {
             .withContext(MT_CONTEXT)
             .build();
         try {
-            BASE_TEST_UTILS.createTenantTables(mtDynamoDb);
+            MtAmazonDynamoDbStreamsBaseTestUtils.createTenantTables(mtDynamoDb);
 
             int i = 0;
-            final MtRecord expected1 = BASE_TEST_UTILS.putTestItem(mtDynamoDb, BASE_TEST_UTILS.TENANTS[0], i++);
-            final MtRecord expected2 = BASE_TEST_UTILS.putTestItem(mtDynamoDb, BASE_TEST_UTILS.TENANTS[0], i);
+            final MtRecord expected1 = MtAmazonDynamoDbStreamsBaseTestUtils
+                .putTestItem(mtDynamoDb, MtAmazonDynamoDbStreamsBaseTestUtils.TENANTS[0], i++);
+            final MtRecord expected2 = MtAmazonDynamoDbStreamsBaseTestUtils
+                .putTestItem(mtDynamoDb, MtAmazonDynamoDbStreamsBaseTestUtils.TENANTS[0], i);
             i = 0;
-            final MtRecord expected3 = BASE_TEST_UTILS.putTestItem(mtDynamoDb, BASE_TEST_UTILS.TENANTS[1], i++);
-            final MtRecord expected4 = BASE_TEST_UTILS.putTestItem(mtDynamoDb, BASE_TEST_UTILS.TENANTS[1], i);
+            final MtRecord expected3 = MtAmazonDynamoDbStreamsBaseTestUtils
+                .putTestItem(mtDynamoDb, MtAmazonDynamoDbStreamsBaseTestUtils.TENANTS[1], i++);
+            final MtRecord expected4 = MtAmazonDynamoDbStreamsBaseTestUtils
+                .putTestItem(mtDynamoDb, MtAmazonDynamoDbStreamsBaseTestUtils.TENANTS[1], i);
 
             // get shard iterators
             MtAmazonDynamoDbStreams mtDynamoDbStreams = MtAmazonDynamoDbStreams.createFromDynamo(mtDynamoDb,
@@ -94,7 +100,7 @@ class MtAmazonDynamoDbStreamsByTableTest {
             // test without context
             final List<Stream> streams = mtDynamoDbStreams.listStreams(new ListStreamsRequest()).getStreams();
             final List<Record> actual = streams.stream()
-                .map(stream -> BASE_TEST_UTILS.getShardIterator(mtDynamoDbStreams, stream))
+                .map(stream -> MtAmazonDynamoDbStreamsBaseTestUtils.getShardIterator(mtDynamoDbStreams, stream))
                 .map(Optional::get)
                 .map(iterator -> mtDynamoDbStreams.getRecords(new GetRecordsRequest().withShardIterator(iterator))
                     .getRecords())
@@ -103,23 +109,28 @@ class MtAmazonDynamoDbStreamsByTableTest {
             assertRecordsEquals(actual, expected1, expected2, expected3, expected4);
 
             // test with tenant contexts
-            MT_CONTEXT.withContext(BASE_TEST_UTILS.TENANTS[0], () -> {
-                String tenantIterator = BASE_TEST_UTILS.getShardIterator(mtDynamoDbStreams, mtDynamoDb).orElseThrow();
-                BASE_TEST_UTILS.assertGetRecords(mtDynamoDbStreams, tenantIterator, expected1, expected2);
+            MT_CONTEXT.withContext(MtAmazonDynamoDbStreamsBaseTestUtils.TENANTS[0], () -> {
+                String tenantIterator = MtAmazonDynamoDbStreamsBaseTestUtils
+                    .getShardIterator(mtDynamoDbStreams, mtDynamoDb).orElseThrow();
+                MtAmazonDynamoDbStreamsBaseTestUtils
+                    .assertGetRecords(mtDynamoDbStreams, tenantIterator, expected1, expected2);
             });
-            MT_CONTEXT.withContext(BASE_TEST_UTILS.TENANTS[1], () -> {
-                String tenantIterator = BASE_TEST_UTILS.getShardIterator(mtDynamoDbStreams, mtDynamoDb).orElseThrow();
-                BASE_TEST_UTILS.assertGetRecords(mtDynamoDbStreams, tenantIterator, expected3, expected4);
+            MT_CONTEXT.withContext(MtAmazonDynamoDbStreamsBaseTestUtils.TENANTS[1], () -> {
+                String tenantIterator = MtAmazonDynamoDbStreamsBaseTestUtils
+                    .getShardIterator(mtDynamoDbStreams, mtDynamoDb).orElseThrow();
+                MtAmazonDynamoDbStreamsBaseTestUtils
+                    .assertGetRecords(mtDynamoDbStreams, tenantIterator, expected3, expected4);
             });
         } finally {
-            BASE_TEST_UTILS.deleteMtTables(mtDynamoDb);
+            MtAmazonDynamoDbStreamsBaseTestUtils.deleteMtTables(mtDynamoDb);
         }
     }
 
     private void assertRecordsEquals(List<Record> actual, MtRecord... expected) {
         assertEquals(expected.length, actual.size());
         for (MtRecord expectedRecord : expected) {
-            assertTrue(actual.stream().anyMatch(actualRecord -> BASE_TEST_UTILS.equals(expectedRecord, actualRecord)));
+            assertTrue(actual.stream()
+                .anyMatch(actualRecord -> MtAmazonDynamoDbStreamsBaseTestUtils.equals(expectedRecord, actualRecord)));
         }
     }